Overhauls Jukebox Sound (#552)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

I spent an absurd amount of time tweaking the jukebox formula and
getting the code to work so please clap.

This PR basically makes it so that the sound Jukeboxes make is more
pleasant by doing the following:

- Jukeboxes can go to 100 volume, but the actual volume that players can
hear is volume capped to 50. The higher volume just affects how far away
the jukebox can be heard from.
- Jukeboxes will play out of both left and right speakers/headphones
unless you're far enough away.
- The maximum distance at which Jukeboxes play is significantly reduced.




## Why It's Good For The Game

Jukeboxes currently have a bug where they are 100 volume by default, and
can be set to that via pressing the reset button. This fixes that and
adds a custom sound system so that music doesn't exclusively play in one
ear if you're to the left or right of a jukebox.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl: BurgerBB
add: Significantly improves Jukebox sound calculations.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->