Hugboxes food germs (#535)
Draft because untested.

## About The Pull Request

Increases disease time to develop on floorfood from 5 seconds to 2
minutes.
The virus you get from floorfood is a pseudo-randomly generated "safe"
virus with a standard virology cure.

## Why It's Good For The Game

I really don't know where to start with this PR so this may seem
scattered.

While this game isn't supposed to be realistic, and eating food on the
floor is not recommended, you (generally) would not get a disease from
eating food that's been on the floor (with this factor alone) for 5
seconds. That's not how immune systems work, and that is not how viruses
work. It also does not make sense mechanic wise considering that
harvesting plants without a botany bag (something most people do) would
almost nearly guarantee a virus. I changed the time from 5 seconds to 5
minutes because 5 minutes is how long it takes for food to go stale when
it's been on the floor.

The virus in question given from floor food nonsense is always a
meme-trier virus. A virus that makes you bite people and the cure is
carpet. A virus that makes you gassy and the cure is firefighting foam.
A virus that makes you have an upset stomach and the cure is...
bluespace crystals. These all seem like good viruses to have when you
interact with a carp, inhale nitryl, or use a teleporter, but not with
eating food off the floor. I tweaked it so that you get a pseudo-random
virus with a "smartly" randomly generated name.

Here is the table of possible symptoms. They've been hand picked based
on my opinion of what would happen to you if you ate dirty food on a
space station.

```
	var/list/possible_symptoms = list( //Not a static list because pick_n_take wouldn't work.
		/datum/symptom/fever,
		/datum/symptom/chills,
		/datum/symptom/dizzy,
		/datum/symptom/headache,
		/datum/symptom/itching,
		/datum/symptom/viraladaptation,
		/datum/symptom/viralevolution,
		/datum/symptom/vomit,
		/datum/symptom/undead_adaptation,
		/datum/symptom/inorganic_adaptation,
		/datum/symptom/heal/metabolism,
		/datum/symptom/heal/plasma
	)
```

## Changelog


:cl: BurgerBB
balance: Increases disease time to develop on floorfood from 5 seconds
to 2 minutes. The disease you get from floorfood is a pseudo-randomly
generated "safe" virus with a standard virology cure.
/:cl: