Nukes the ability to buy most ballistic weapons from cargo. (correct this time) (#486)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Removes the ability to buy many ballistic weapons from cargo, both as
crates and as goodie packs.

last time I tried to merge from skrat instead of bubber.. whoops!
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Why It's Good For The Game
Incredibly unfun to fight against incredibly OP ballistic weapons like
this. Skyrat thought it was a 'good and balanced change', well, it
isn't. This is why we removed most of the ballistic weapons from the
armory in the first place. This is a server with antags that sometimes
want to have a chance. No completely demolishing them with your
shotgun-pistol bloodloss-embed machine instantly.
Some weapons, like riot shotguns, MCRS, the plasma thrower, the
detective's gun, and the pepperball, stay because they fit more in line
with the game.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
del: Removed the ability to buy most ballistic weapons from guncargo.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: oilysnake <63020759+oilysnake@users.noreply.github.com>
Co-authored-by: nevimer <77420409+nevimer@users.noreply.github.com>