Improves Harm Alarm cosmetic code and adds customization. (#541)
## About The Pull Request

Harm alarm is now improved by changing the sound played, the default
message, and the ability to change the message.



https://github.com/Bubberstation/Bubberstation/assets/8602857/32430605-63cf-4c9e-97bb-8ec992dbd81f




## Why It's Good For The Game

"Human Harm" is outdated given that everyone is a white haired catgirl
woman.

## Changelog

:cl: BurgerBB
add: Harm alarm is now improved by changing the sound played, the
default message, and the ability to change the message.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: nevimer <77420409+nevimer@users.noreply.github.com>