Fixes Air Alarms resetting scrubbers/vents on map load. (#550)
## About The Pull Request

What it says on the can. Fixes Air Alarms resetting scrubbers/vents on
map load. What was happening is that air alarms were forcing
scrubbers/vents to go to their default setting, which is a problem for
vents that siphon or scrubbers that scrub out certain things.

This should be testmerged considering things may or may not go wrong
with very specific and really niche issues such as for some reason a far
off ruin that hasn't been worked on since 2014 somehow relying on this
bug to exist to have bad air.

Will fix the bug upstream on /tg/ once there are no issues found.

## Why It's Good For The Game

Fixes are good.

## Changelog

:cl: BurgerBB
fix: Fixes Air Alarms resetting scrubbers/vents on map load.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->