PR for donator items (#582)
Inclusion for the following sprites:

- Donorequest-0015: Bubbledragon plushie for Astroturf in plushes.dmi

- Donorequest-0017: Adds items for Catamin - ID card, wine bottle and a
surgery kit along with all tools for it in a new folder for donator
items.

- Donorequest-0005: Adds items for Nikolai "Rabid" Volkov - clothing, ID
card and a plushie.

- Donorequest-0018: Adds new body markings for digitigrade legs for owl
people and a new set of wings with *aflap emote and owl beak for snouts.

TO DO: Implementation for the code of the items. Please message me at
keila in discord if you have questions about the implementation of these
items since some have special properties and different locations in the
modular_zubbers folder.