Added *chirp, modified *caw (#579)
## About The Pull Request

Added *chirp cuz it's essential for a birdier communication. Also
followed the same structure for emotes as modular skyrat so more custom
emotes can be added there in the future without it being messy!

And yes, I've tested it and it works, woo!

## Why It's Good For The Game

Birds are good. Bird communication is also good.

## Changelog

:cl: Caw!
add: Added *chirp, modified *caw and *caw2
/:cl: