Tweaks Validmov (Safeguard) to something less validhunty and more in tune with the new rules. (#546)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Changes Safeguard from:

> "Safeguard: Protect your assigned space station and its assets without
unduly endangering its crew.",
> "Prioritize and Comply: The directives and safety of crew members are
to be prioritized according to their rank, role, and need, while
preserving their safety and well-being.",
> "Survive: You are not expendable. Do not allow unauthorized personnel
to tamper with or damage your equipment",

to:

>"Safeguard: Protect your assigned station and assets without
endangering its crew.",
>"Survive: Do not allow unauthorized personnel to tamper with or damage
your equipment.",
>"Comply: The directives of crew members are to be prioritized according
to their rank and need."

Renames the lawset from "Armadyne OS Safeguard V1.0" to "NT OS Safeguard
V1.0"


## Why It's Good For The Game


### Law 1: Safeguard
> OLD: "Safeguard: Protect your assigned space station and its assets
without unduly endangering its crew."
> NEW: "Safeguard: Maintain your assigned station and assets without
endangering its crew."

- Removed space in space station because icebox exists,
- Removed the prefix "unduly" because the law basically meant "you can
harm crew if they deserve it"
- Changed "protect" to "maintain" as maintain is more of a softer phrase
and it does not encourage active threat seeking.

### Law 2: Comply
> OLD: "Prioritize and Comply: The directives and safety of crew members
are to be prioritized according to their rank, role, and need, while
preserving their safety and well-being."
> NEW: "Comply: Follow and prioritize the directives of the station's
crew members according to their rank, role, and need."

- Removed "Prioritize" because I want to keep the 1 word verb theme.
- Pretty much rewrites this law to make it more clear that you are to
follow orders.
- The "preserving their safety and well-being" part is redundant because
that part is handled via law 1.

### Law 3: Survive
> OLD: "Survive: You are not expendable. Do not allow unauthorized
personnel to tamper with or damage your equipment"
> NEW: "Survive: Do not allow unauthorized personnel to tamper with or
damage your equipment."

- Removed the "You are not expendable." part because it was redundant.

### Overall

It is my personal opinion that these changes are needed because
Safeguard was a super validhunt heavy lawset and allowed borgs to play
security 2.0. This was by design given the fact that it literally has
"Armadyne" in the module title and I believe that the creators wanted a
more militarized station with the security overhaul given that that
entire change was inspired by Eris. Given how Bubberstation is straying
away from that, I think we should do away with the current default
lawset to something that encourages AI and Cyborg players to preserve
life while still being reasonable for something owned by a Corporation.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl: BurgerBB
balance: Changes the Safeguard lawset to something less validhunty.
/:cl: