Fixes Burgerstation Disposals, Detective Access, and Bitrunner access. (#586)
## About The Pull Request

Disposals is fixed. Unfortunately because of a code issue, mail does not
work when put into the conveyor, but it does work when put into the
newly placed disposals container in the mail room.

Access for bitrunners and detectives are fixed.




## Changelog

:cl: BurgerBB
fix: Fixes Burgerstation Disposals, Detective Access, and Bitrunner
access.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->