Subverts Democracy and destroys the Meta-Icebox-Delta voting cycle by adding negative voting multipliers to recently played maps. (#554)
Draft because controversial.

## About The Pull Request

Your vote for a map only counters 75% if the vote was a recently played
map.
The last 4 rounds are considered recently played maps (not the last 4
maps, which means that if the cycle (from most recent to least recent)
was "icebox, delta, delta, meta, biodome", it will only count icebox,
delta, meta.
If a map was voted for twice in the last 4 rounds, it cannot be chosen.

## Why It's Good For The Game

It sucks seeing server status and the map rotation is just:

ICEBOX, DELTA, META, ICEBOX, DELTA, META, TRAM, ICEBOX, DELTA, META,
ICEBOX, DELTA, META, TRAM with the occasional biodome, limastation, or
burgerstation map vote because it's lowpop and it seems only lowpop
people like those maps.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl: BurgerBB
add: Your vote for a map only counters 75% if the vote was a recently
played map (recently refers to the last 4 rounds).
add: If a map was voted for twice in the last 4 rounds, it cannot be
chosen.
/:cl: