Fixes HoS/Warden drip, disables HoS beacon (#629)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds some of the Skyrat Removed HoS and warden clothing to their garment
bags, because drip is based
Moved the choice beacon to a modular file
Disabled the choice beacon from spawning because it doesn't work

## Why It's Good For The Game

Red is back in fashion, its time to bring it back for good.
All the guns in the HoS beacon were missing/broken so I disabled it
until someone fixes it lol

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: HoS and Warden red drip
del: HoS choice beacon, because none of the guns were actually
spawning/functional
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->