Re-add and update Kilostation (#614)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds Kilostation back in, fully reverting it's deletion and adding it in
as a bubber map.

I've fully brought Kilo up to date and added in the coroners office and
bitrunner's den as best as I could, as well as adding a smattering of
those fancy new digital clocks, and bandaids. As far as I can tell it's
ready to go.

And I figure 75 is a good max player count for Kilo, it **is** pretty
cramped, could do 80 though maybe.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Why It's Good For The Game
Kilostation is an amazing lowpop map and is well loved, and didn't
deserve to be killed off.
This will add a new map to use during low-pop other than birdshot and
voidraptor.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog
:cl:
add: Re-added Kilostation
add: Bitrunner Den and Coroners office crammed into Kilostation
fix: More things than I care to list to get Kilo up to date
fix: Aligned Kilo's arrival dock with both shuttle doors.
config: Made Kilostation voteable up to a max player count of 75 by
default
/:cl:
<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->


<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->