[FULLY MODULAR] New quirk: Hungry (#604)
## About The Pull Request

Adds a new neutral quirk, hungry, with the current effect of doubling
the decay of one's nutrition.
Neutral because Moffy told me to make it so, might make it a -1 or add
something positive to it if others have a different opinion.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Why It's Good For The Game

Even more reason to interact with the kitchen for those who take this
quirk. Please take it I need someone to talk to


![image](https://github.com/Bubberstation/Bubberstation/assets/49160555/01e1ef13-f425-49ed-9fbb-81e3b387d824)

This is not a quirk for those fat fetishists to abuse like some of you
claimed, they already have their one

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: New quirk: Hungry
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->