Adds contribution rules to the main readme.txt (#682)
I'm making rules and they will be followed when this is merged.

## Contribution Rules and Guidelines

**1. Do Not Be A Dick**
- The Bubberstation main repository is run by and contributed by
volunteers. You are not entitled to our time and energy. We reserve the
right to permanently remove anyone who does not show both our
contributor's and maintainer's common decency.

**2. This repository will only accept feedback that has to do with the
functionality of code, sound, and images**
- To maintain a proper working space for our contributors and to keep
the repository free of clutter. All feedback that does not have to do
with the functionality, quality, and review of code will be marked
off-topic.
- If you feel like feedback on a controversial PR is required, you are
free to open a [feature
request](.https://github.com/Bubberstation/Bubberstation/issues/new/choose)
under the repository issues page and reference the PR number.

**3. The Licensing is Non-negotiable**
- You are free to take, redistribute, modify, and readapt any code or
commit found on this repository.
- All code files are under **GNU AGPL V3** 
- All asset files (images and sound) are **CC-BY-SA 3.0** unless
otherwise stated
- The license information, including the MIT license and its exceptions,
can be found at the bottom of this readme.
- What does this mean to our contributors? **The** [GNU AGPL
V3](.https://www.gnu.org/licenses/agpl-3.0.html) **licensing on this
codebase is non-negotiable** and **irrevocable** the moment you open a
PR. We are incapable of privatizing this codebase and you are incapable
of restricting your code contributions even if you close the PR. Do not
open a contribution if you do not feel comfortable with your code being
permanent across the entirety of the community. **All contributions are
preserved in commit history**
- What does this mean to our spriters and sound designers? [CC-BY-SA
3.0](.https://creativecommons.org/licenses/by-sa/3.0/) **is the only
license that this project will accept** unless given express permission
by the owners of this repository.
- It is prohibited to use the attribution requirement to [suggest that
you endorse or support a particular
use](.https://creativecommons.org/faq/#do-i-need-to-be-aware-of-anything-else-when-providing-attribution)
of your assets.
- You are entitled to credit yourself with comments and you are entitled
to waive the attribution requirement choosing not to be identified as
the creator if you wish.
- If you do not like how your assets were modified or used, it is
required that the other person [remove the attribution information upon
request](.https://wiki.creativecommons.org/wiki/License_Versions#Licensors_may_request_removal_of_attribution).
- Modifications or adaptions must disclose the source, the author, and
[any changes you've
made](.https://wiki.creativecommons.org/wiki/License_Versions#Modifications_and_adaptations_must_be_indicated).
- The goon code is incompatible with this codebase and will not be
accepted.