Ports Gimmick Traitor from Daedalus Dock (and raises starting TC) (#583)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

![mstsc_mBeessMCFA](https://github.com/Bubberstation/Bubberstation/assets/77420409/af879190-1f75-4c63-a4eb-aaa35d29cc5b)

![mstsc_821RI1VnPn](https://github.com/Bubberstation/Bubberstation/assets/77420409/51d81ffb-5962-4ee4-81a2-e9e4652c6e70)

ports the meat from
https://github.com/DaedalusDock/daedalusdock/pull/564

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Why It's Good For The Game
I don't think anyone but the player of traitors enjoys progressive
traitor. This replaces it with the gimmick traitor goals, which are much
more entertaining for everyone. Big QOL.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Traitor Most traitor goals have been replaced with gimmick traitor
goals.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: iero-srv <>
Co-authored-by: projectkepler-RU <99981766+projectkepler-ru@users.noreply.github.com>
Co-authored-by: The Sharkening <95130227+StrangeWeirdKitten@users.noreply.github.com>