[Ready] Firelock (and air alarm) Overhaul (#492)
## About The Pull Request

Changes how firelocks automatically trigger. Firelocks are handled
completely by the air alarm, and don't have their own atmos checks.

Reverts firelock crowbar behavior back to /tg/'s behavior.

Actually makes it so that crowbars stay open when forced open via right
click.

Firelocks no longer automatically close when forced open via manual
override (alt click). Reduces manual override time from 10 seconds to 5
seconds.

Changes firelock sprites to be less visible when not active.

Air alarms slowly heat the room if the temperature is too low.

## Why It's Good For The Game

Firelocks are some of the worst and most annoying things to deal with in
the game. They're incredibly inconsistent in terms of how they actually
function, and can be a downright nuisance to deal with. This PR makes it
so that firelocks are controlled by the area's air alarm status, thus
eliminating any confusion as to why the shutters keep fucking closing
for no real reason.

I recommend a testmerge for this so feedback can be given about the new
system given this functionally alters the power of breaches. As a
consequence of this PR, average case scenario is that adjacent rooms
will run into atmos issues as well, but they will be much easier to
repair and diagnose.

## Changelog

:cl: BurgerBB
add: Firelocks are triggered only by air alarms (or manual input).
Reverts firelocks back to /tg/'s behavior, and fixes a lot of bugs. The
description for examining firelocks is now actually accurate!
add: Air alarms slowly heat the room if the temperature is too low.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: nevimer <77420409+nevimer@users.noreply.github.com>
Co-authored-by: The Sharkening <95130227+StrangeWeirdKitten@users.noreply.github.com>