ss13_construct optimization pull request (#711)
## About The Pull Request

This gradually decreases the amount of varedited objects on the map,
turning most of them into subtypes. I guess I should've made that
before.


## Why It's Good For The Game

Being able to actually pass the tests seems to be a good idea.

## Changelog


:cl:
refactor: changed most of the varedits on ss13_construct into subtypes
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: Resomi Enjoer <fenyshaarmagedon.gmai.com>
Co-authored-by: Fenysha <126053988+TeshariEnjoer@users.noreply.github.com>
Co-authored-by: nevimer <77420409+nevimer@users.noreply.github.com>